requirements.txt
acres==0.2.0
aiofiles==24.1.0
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.12.0
antlr4-tools==0.2
anyio==4.8.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
backports.tarfile==1.2.0
beautifulsoup4==4.12.0
bitarray==3.0.0
blinker==1.4
bs4==0.0.1
cbor2==5.4.6
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==3.1.0
ci-info==0.3.0
ckzg==1.0.2
click==8.1.8
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
contourpy==1.3.1
cryptography==3.4.8
crytic-compile==0.3.7
cycler==0.12.1
cytoolz==1.0.1
dbus-python==1.2.18
decorator==4.4.2
distro==1.7.0
etelemetry==0.3.1
eth-account==0.11.3
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==1.0.1
eth-typing==4.4.0
eth-utils==4.1.1
eth_abi==5.1.0
exceptiongroup==1.2.2
filelock==3.16.1
fonttools==4.55.8
frozenlist==1.3.3
graphviz==0.20.3
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
idna==3.4
importlib_metadata==8.4.0
importlib_resources==6.5.2
install-jdk==1.0.2
isodate==0.6.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.6.0
kiwisolver==1.4.8
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
looseversion==1.3.0
lru-dict==1.2.0
lxml==4.9.2
markdown-it-py==3.0.0
matplotlib==3.10.0
mdurl==0.1.2
more-itertools==10.5.0
multidict==6.0.4
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==2.2.1
oauthlib==3.2.0
openai==1.59.6
packaging==23.0
panda==0.3.1
pandas==2.2.3
parsimonious==0.10.0
pathlib==1.0.1
pillow==11.1.0
prettytable==3.12.0
protobuf==5.29.3
prov==2.0.1
puremagic==1.28
pycodestyle==2.10.0
pycparser==2.22
pycryptodome==3.17
pydantic==2.10.5
pydantic_core==2.27.2
pydot==3.0.4
Pygments==2.18.0
pygraphviz==1.14
PyJWT==2.3.0
pyparsing==3.2.1
pysha3==1.0.2
python-dateutil==2.9.0.post0
pytz==2024.2
pyunormalize==16.0.0
pyxnat==1.6.2
PyYAML==6.0
rdflib==6.3.2
referencing==0.35.1
regex==2024.11.6
requests==2.28.2
rich==13.7.1
rlp==4.0.1
rpds-py==0.22.3
safe-pysha3==1.0.4
scipy==1.15.1
SecretStorage==3.3.3
simplejson==3.19.3
six==1.16.0
slither-analyzer==0.10.4
sniffio==1.3.1
solc-select==1.0.4
soupsieve==2.4
starlette==0.45.2
tiktoken==0.4.0
tomli==2.0.1
toolz==1.0.0
tqdm==4.65.0
traits==6.4.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.15
uvicorn==0.34.0
wadllib==1.3.6
wcwidth==0.2.6
web3==6.20.3
websockets==14.1
yarl==1.8.2
z3-solver==4.11.2.0
zipp==1.0.0
